Add `describe` function to SPK and BPC
Will add it to Almanac
