+++ anise-cli/Cargo.toml
@@ -12,7 +12,6 @@ description = "A command line interface for ANISE"
 anise = { workspace = true }
 clap = { version = "4", features = ["derive"] }
 pretty_env_logger = { workspace = true }
-tabled = { workspace = true }
 bytes = { workspace = true }
 memmap2 = { workspace = true }
 snafu = { workspace = true }
