+++ anise/Cargo.toml
@@ -32,6 +32,7 @@ serde_derive = { workspace = true }
 serde_dhall = { workspace = true, optional = true }
 reqwest = { version = "0.11.23", optional = true, features = ["blocking"] }
 platform-dirs = { version = "0.3.0", optional = true }
+tabled = { workspace = true }
 
 [dev-dependencies]
 rust-spice = "0.7.6"
